backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cassandra-reddit-mcp"
version = "0.1.0"
description = "Reddit MCP server — subreddit browsing, search, post/comment viewing"
requires-python = ">=3.12"
dependencies = [
    "cassandra-mcp-auth @ git+https://github.com/Cassandras-Edge/cassandra-auth.git@53f5f18#subdirectory=python",
    "cryptography>=44.0.0",
    "fastmcp[auth]>=2.12.0",
    "httpx>=0.28.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
cassandra-reddit-mcp = "cassandra_reddit_mcp.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
